able to create new chat and completed view supplier and product feature

+++ P2_SubChain/Models/Product.cs
@@ -12,14 +12,13 @@ namespace P2_SubChain.Models
 
         public int UserId { get; set; }
 
-
         public string ProductDesc { get; set; }
 
         public string ProductName { get; set; }
 
         public string ProductType { get; set; }
 
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
 
         public string ImageUrl { get; set; }
